requirements.txt
flask
pillow
backgroundremover